Rename types.ts -> spørsmål.ts
