+++ src/frontend/components/SøknadsSteg/OmDeg/spørsmål.ts
